+++ software/rocketlogger/sensor/tsl4531.c
@@ -68,7 +68,7 @@ int tsl4531_init(int sensor_identifier) {
         return FAILURE;
     }
 
-    uint8_t sensor_id = tsl4531_get_id();
+    int sensor_id = tsl4531_get_id();
     if (sensor_id != TSL4531_ID) {
         rl_log(ERROR, "TSL4531 with wrong sensor ID: %d", sensor_id);
         return FAILURE;
